% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{adapt_tuner_double_exp}
\alias{adapt_tuner_double_exp}
\title{Default proposal adaptation}
\usage{
adapt_tuner_double_exp(term, acc_counter, prop_sd, interval_length)
}
\arguments{
\item{term}{a character string for name of parameter to tune within acc_counter}

\item{acc_counter}{the list object with acceptance counts}

\item{prop_sd}{list of proposal sds}

\item{interval_length}{length of adaptation interval (default should be 100)}
}
\value{
Updated proposal SD for this term
}
\description{
By default, prop_sd will be updated for MH blocks based on a double exponential
targetting a 0.17 to 0.5 acceptance rate. Note this current set up allows for
bulk updates of multiple parameters if any fall outside the goal range
}
\keyword{internal}
