% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-engines.R
\name{adaptive_mcmc_run}
\alias{adaptive_mcmc_run}
\title{MCMC Run}
\usage{
adaptive_mcmc_run(
  data_list,
  const_list,
  init_list,
  prior_pars,
  update_functions,
  prop_sd = NULL,
  save_names = NULL,
  derv_quants = NULL,
  derived_functions = NULL,
  adapt_tuner = adapt_tuner_double_exp,
  chain_check = geweke_check,
  min_burn = 5000,
  max_iter = 1e+06,
  adapt_intv = 100,
  max_thin = 10,
  goal_ess = 100
)
}
\arguments{
\item{min_burn}{integer of minimum number of burnin interations before saving}

\item{max_iter}{the maximum number of iterations}

\item{adapt_intv}{adaptation interval for both proposal tunning and thinnin tuning}

\item{max_thin}{maximum thinning interval}

\item{goal_ess}{target ess for a chain}
}
\value{
a list of sample objects and info.
}
\description{
A generic mcmc engine for flexible model building
}
\details{
This function is the main driver of mcmc algos in a base R framework,
it is designed for maximum flexibility in mind, not necessarily speed.
However performance largely depends

There are several useful features including an adapative-burnin, thinning, and tuning
for metropolis hastings algorithms. Users should carefully note that the burnin and thinning
adaptation algorithms rely on parameters which are saved ONLY. Thus, if you are running a complex
model with multiple variables but only saving a few, there may be hidden issues.
Also be mindful of default max/min args, these can create a big memory constraint if
saving many paramters
}
\keyword{internal}
