% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-handler.R
\name{adjust_divisor}
\alias{adjust_divisor}
\title{Adjust inputs}
\usage{
adjust_divisor(x, v)
}
\arguments{
\item{x}{larger val}

\item{v}{smaller val}
}
\value{
adjusted value of y
}
\description{
Adjust inputs
}
\keyword{internal}
