% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{afas}
\alias{afas}
\alias{aafa}
\alias{gafa}
\alias{fafa}
\title{Along First Axis functions}
\usage{
aafa(arr, idx, val)

gafa(arr, idx)

fafa(arr)
}
\arguments{
\item{arr}{array or vector}

\item{idx}{index to get on first axis}

\item{val}{value to assign at index}
}
\value{
array with changes at idx
}
\description{
These series of helper functions provide an extremely
valuable approach to interact with the first axis of any shaped array.
For MCMC applications where the first axis is iterations of a saved value.
A- assign (replace values); G- get (index values); F- flatten (return set of chains)
}
