% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assess_burnin}
\alias{assess_burnin}
\title{Assess Burnin}
\usage{
assess_burnin(save_obj, chain_check)
}
\arguments{
\item{save_obj}{a list of save objects that IS TRIMMED!}

\item{chain_check}{a function for checking chains - default should be geweke_check}
}
\value{
boolean if it passess the checker
}
\description{
Loop through the save object to assess if the burnin meets
some convergence criteria. Default is to use geweke_check but other
functions can be created.
}
\keyword{internal}
