% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-engines.R
\name{check_function_format}
\alias{check_function_format}
\title{Check update and derived function formats}
\usage{
check_function_format(fn)
}
\arguments{
\item{fn}{function to check}
}
\description{
This makes sure that the update/derived functions are properly formatted
and will return a list with val and acc (if MH).
Also it will check formatting between val names and prop_sd names (if needed).
}
\keyword{internal}
