% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-signal-process.R
\name{curve-signal-process}
\alias{curve-signal-process}
\alias{mc_get_tm}
\alias{mc_rm_trend}
\alias{detect_tm}
\alias{mc_baseline}
\title{Detect peaks in Melting Curve and return TM values}
\usage{
mc_get_tm(
  mc,
  npeaks = 3,
  threshold = 0,
  sortstr = TRUE,
  zero = "+",
  well_position = NULL
)

mc_rm_trend(mc)

detect_tm(data, ...)

mc_baseline(mc, method = c("spline", "polyfit"), degree = 3)
}
\arguments{
\item{mc}{a MeltingCurve object}

\item{npeaks}{the number of peaks to return}

\item{threshold}{the minimum }

\item{sortstr}{logical; should the peaks be returned sorted in decreasing
        oreder of their maximum value}

\item{zero}{can be `+', `-', or `0'; how to interprete succeeding steps
              of the same value: increasing, decreasing, or special}

\item{well_position}{if NULL return all wells}

\item{data}{a tibble with \code{temperature} and \code{derivative} values for melting curve}

\item{...}{will pass to \code{pracma::findpeaks()}}

\item{method}{c("spline","polyfit")}

\item{degree}{degree of freedom}
}
\value{
a tibble

a processed Melting Curve object
}
\description{
Detect peaks in Melting Curve and return TM values

Remove trend in a melting curve

Remove the baseline of melting curve
}
\seealso{
\code{\link[pracma:findpeaks]{pracma::findpeaks()}}
}
