% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-interpolation.R
\name{curve_resample}
\alias{curve_resample}
\title{Re-sample melting curve data}
\usage{
curve_resample(
  curve_data,
  group = "well_position",
  from = 70,
  to = 90,
  by = 0.1,
  variable = c("fluorescence", "derivative"),
  method = c("linear", "constant", "nearest", "spline", "pchip", "cubic")
)
}
\arguments{
\item{curve_data}{raw data}

\item{group}{group}

\item{from}{start of temperature}

\item{to}{end of temperature}

\item{by}{step}

\item{variable}{c("fluorescence", "derivative")}

\item{method}{c("linear", "constant", "nearest", "spline", "pchip", "cubic")}
}
\value{
a tibble
}
\description{
Re-sample melting curve data
}
