% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{find_file}
\alias{find_file}
\title{Load Recent helper:}
\usage{
find_file(chk_path, files)
}
\arguments{
\item{chk_path}{the path to look for file}

\item{files}{the file list to try and readfrom}
}
\value{
file path
}
\description{
Load Recent helper:
}
\keyword{internal}
