% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-engines.R
\name{format_output}
\alias{format_output}
\title{format outputs for mcmc algos}
\usage{
format_output(save_item, derv_save, drop_front = TRUE)
}
\arguments{
\item{save_item}{save object as a list with named parameters}

\item{derv_save}{derv save object as list with named parameters OR NULL}
}
\value{
formatted output for mcmc engines
}
\description{
format outputs for mcmc algos
}
\keyword{internal}
