% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{load_all_blks}
\alias{load_all_blks}
\title{Reconstruct states}
\usage{
load_all_blks(state)
}
\arguments{
\item{state}{path to check}
}
\value{
completed set of chains
}
\description{
Reconstruct states
}
\keyword{internal}
