% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{load_recent}
\alias{load_recent}
\title{Find Most Recent Save Point}
\usage{
load_recent(chk_path)
}
\arguments{
\item{chk_path}{path to look in}
}
\value{
mcmodel state objs
}
\description{
Find Most Recent Save Point
}
\keyword{internal}
