% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{make_save_block}
\alias{make_save_block}
\title{Constructor for checkpoint paths}
\usage{
make_save_block(init_list, save_names, alloc)
}
\arguments{
\item{init_list}{a list for initialization}

\item{save_names}{pars to keep track of}

\item{alloc}{how big each block needs to be}
}
\value{
block list strucutre
}
\description{
Constructor for checkpoint paths
}
\keyword{internal}
