% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{make_state}
\alias{make_state}
\title{Construct a new state for mcmc}
\usage{
make_state(
  data_list,
  const_list,
  init_list,
  prior_pars,
  prop_sd,
  update_functions,
  derived_functions,
  save_names,
  derv_quants,
  n_iter,
  thin,
  burn_adapt,
  min_burn,
  n_burn,
  prop_adapt,
  adapt_tuner,
  chain_check,
  chkpt_path,
  chkpt_freq
)
}
\arguments{
\item{data_list}{list object with non-changing data}

\item{const_list}{list object with relevant indexing used in updates/derived fns}

\item{init_list}{list object with ALL parameters used in update/derived, see \link{mcmc_structures}}

\item{prior_pars}{list of prior values used in update functions}

\item{prop_sd}{IF any Metropolis-Hastings algos are used with, prop_sd much be provided.}

\item{update_functions}{a list of update functions used in blocks for MCMC application}

\item{derived_functions}{List of functions for computing derived quantities.}

\item{save_names}{(optional) list of parameters in init_list to save out. Defaults to init_list}

\item{derv_quants}{Character vector specifying derived quantities to track.}

\item{n_iter}{integer for number of interations}

\item{thin}{integer for thinning interval. Default is 1.}

\item{burn_adapt}{logical to decide to use adaptive burn-in procedure.}

\item{min_burn}{integer used if burn_adapt=TRUE, how long minimum burn in needs to be}

\item{n_burn}{integer for maximum burn-in length}

\item{prop_adapt}{interval for tuning proposal in a MH-example}

\item{chkpt_freq}{interval of when to write out blocked chunks}
}
\value{
state object
}
\description{
McModel States are the core interal object for mcmc runs
They allow for checkpoint (chkpt) saving which can allow for in-memory action
}
\keyword{internal}
