% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting-curve-class.R
\name{mc_pca}
\alias{mc_pca}
\title{PCA analysis of MeltingCurve object}
\usage{
mc_pca(object, scale = FALSE)
}
\arguments{
\item{object}{MeltingCurve object}

\item{scale}{Scale species to unit variance (like correlations).}
}
\value{
PCA result
}
\description{
PCA analysis of MeltingCurve object
}
