% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting-curve-class.R
\name{mc_pca_plot}
\alias{mc_pca_plot}
\title{plot PCA result with extra data (Plate/Sample table)}
\usage{
mc_pca_plot(
  pca,
  extra = NULL,
  color = NULL,
  shape = NULL,
  label = "well_position",
  show_temperature = FALSE,
  temp_n = 3
)
}
\arguments{
\item{pca}{pca result}

\item{extra}{extra data table (join by 'well_position')}

\item{color}{color mapping}

\item{shape}{shape mapping}

\item{label}{label by 'well_position' by default}

\item{show_temperature}{show temperature arrow}

\item{temp_n}{limit for temperature arrows}
}
\value{
ggplot
}
\description{
plot PCA result with extra data (Plate/Sample table)
}
