% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting-curve-class.R
\name{mc_tbl2wider}
\alias{mc_tbl2wider}
\title{Tibble to wider matrix-like format}
\usage{
mc_tbl2wider(
  mc,
  column_names_from = "temperature",
  column_values_from = "derivative"
)
}
\arguments{
\item{mc}{object}

\item{column_names_from}{column names}

\item{column_values_from}{column values}
}
\value{
a wider tibble
}
\description{
Tibble to wider matrix-like format
}
