% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-engines.R
\name{mcmc_run_internal}
\alias{mcmc_run_internal}
\title{Internal dispatch for a single chain run}
\usage{
mcmc_run_internal(
  data_list,
  const_list,
  init_list,
  prior_pars,
  prop_sd,
  update_functions,
  save_names,
  derv_quants,
  derived_functions,
  n_iter,
  thin,
  n_burn,
  burn_adapt,
  min_burn,
  prop_adapt = 100,
  adapt_tuner = adapt_tuner_double_exp,
  chain_check = geweke_check
)
}
\arguments{
\item{data_list}{list object with non-changing data}

\item{const_list}{list object with relevant indexing used in updates/derived fns}

\item{init_list}{list object with ALL parameters used in update/derived, see \link{mcmc_structures}}

\item{prior_pars}{list of prior values used in update functions}

\item{prop_sd}{IF any Metropolis-Hastings algos are used with, prop_sd much be provided.}

\item{update_functions}{a list of update functions used in blocks for MCMC application}

\item{save_names}{(optional) list of parameters in init_list to save out. Defaults to init_list}

\item{derv_quants}{Character vector specifying derived quantities to track.}

\item{derived_functions}{List of functions for computing derived quantities.}

\item{n_iter}{integer for number of interations}

\item{thin}{integer for thinning interval. Default is 1.}

\item{n_burn}{integer for maximum burn-in length}

\item{burn_adapt}{logical to decide to use adaptive burn-in procedure.}

\item{min_burn}{integer used if burn_adapt=TRUE, how long minimum burn in needs to be}

\item{prop_adapt}{integer interval for proposal adaptation. Default is 100}

\item{adapt_tuner}{function for proposal adaptation. Default is \link{adapt_tuner_double_exp}. no_adapt is also a built-in option}

\item{chain_check}{function to check chain convergence within adaptive burnin feaure.}
}
\value{
list object with mcmc features
}
\description{
Internal dispatch for a single chain run
}
\keyword{internal}
