% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmodel-package.R
\docType{package}
\name{mcmodel-package}
\alias{mcmodel}
\alias{mcmodel-package}
\title{McModel}
\description{
Tools for constructing Gibbs-MH samplers in base R with adaptive
MCMC algorithm support and basic diagnostics.
}
\author{
\strong{Maintainer}: Alex J. Barth \email{alex.barth.science@gmail.com}

}
\keyword{internal}
