% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting-curve-class.R
\docType{class}
\name{MeltingCurve-Class}
\alias{MeltingCurve-Class}
\alias{quantstudio2mc}
\title{Class for storing melting curve data}
\usage{
quantstudio2mc(
  x,
  experiment_date = get_quantstudio_run_time(x),
  plate = NULL,
  primer = NULL,
  col_data = c("well_position", "temperature", "derivative")
)
}
\arguments{
\item{x}{a quantstudio result file or object}

\item{experiment_date}{default is \code{Sys.Date()}}

\item{plate}{default is "NA"}

\item{primer}{default is "NA"}

\item{col_data}{keep necessary columns}
}
\value{
a MeltingCurve object
}
\description{
\itemize{
\item \code{MeltingCurve} is a S4 class with several slots.
\item Use \code{quantstudio2mc()} to convert a quantstudio result to \code{MeltingCurve} object.
}
}
\section{Slots}{

\describe{
\item{\code{experiment_date}}{experiment date}

\item{\code{plate}}{plate setting}

\item{\code{primer}}{primer name}

\item{\code{data}}{curve data}
}}

