% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting-curve-class.R
\name{mc2tbl}
\alias{mc2tbl}
\title{Transform a MeltingCurve object to tibble}
\usage{
mc2tbl(mc)
}
\arguments{
\item{mc}{a MeltingCurve object}
}
\description{
Transform a MeltingCurve object to tibble
}
