% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{no_adapt}
\alias{no_adapt}
\title{No adaptation in proposal tuning}
\usage{
no_adapt(term, acc_counter, prop_sd, interval_length)
}
\arguments{
\item{term}{a character string for name of parameter to tune within acc_counter}

\item{acc_counter}{the list object with acceptance counts}

\item{prop_sd}{list of proposal sds}

\item{interval_length}{length of adaptation interval (default should be 100)}
}
\value{
original proposal sd for this term
}
\description{
This is just to keep code smooth but allow for no_adaptation
}
\keyword{internal}
