% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_384.R
\name{plot_384_single_concentration}
\alias{plot_384_single_concentration}
\title{Plot 384 well species quantity of a single species}
\usage{
plot_384_single_concentration(data, species, well_size = 5)
}
\arguments{
\item{data}{definition of 384 plate}

\item{species}{species name}

\item{well_size}{well size}
}
\value{
a ggplot object
}
\description{
Plot 384 well species quantity of a single species
}
