% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting-curve-class.R
\name{plot_mc}
\alias{plot_mc}
\title{Plot a MeltingCurve object}
\usage{
plot_mc(mc, y = "derivative", show_tm = FALSE)
}
\arguments{
\item{mc}{MeltingCurve object}

\item{y}{y axis mapping}

\item{show_tm}{whether show TM value in melting curve plot}
}
\value{
ggplot object
}
\description{
Plot a MeltingCurve object
}
