% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_viz.R
\name{plot_post_chain}
\alias{plot_post_chain}
\title{Plot posterior chain trace plots}
\usage{
plot_post_chain(samples, par_name, real_par = NULL)
}
\arguments{
\item{samples}{a list of named samples}

\item{par_name}{the name to access in samples}

\item{real_par}{if known, a constant to plot (typically only for simulation cases)}
}
\value{
produces a plot in active window
}
\description{
Plot posterior chain trace plots
}
\examples{
\dontrun{
mcmc_out = mcmc_run(...)

plot_post_chain(mcmc_out$samples, 'parname')
}

}
