% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_viz.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot posterior chain trace plots}
\usage{
plot_trace(run_obj, par_name, h_lines = NULL)
}
\arguments{
\item{run_obj}{a list of named samples}

\item{par_name}{the name to access in samples}

\item{h_lines}{if known, a constant to plot across traces (typically only for simulation cases)}
}
\value{
plot in active window
}
\description{
Plot posterior chain trace plots
}
\examples{
\dontrun{
mcmc_out = mcmc_run(...)

plot_post_chain(mcmc_out$samples, 'parname')
}

}
