% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_viz.R
\name{post_summary}
\alias{post_summary}
\title{Posterior Summary}
\usage{
post_summary(run_obj)
}
\arguments{
\item{run_obj}{output from \link{mcmc_run}}
}
\value{
a matrix of summary values
}
\description{
Posterior Summary
}
\examples{
\dontrun{
 mcmc_out = run_mcmc(...)
 post_summary(run_obj)
}

}
