% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{print.mcmod.state}
\alias{print.mcmod.state}
\title{Interal print method}
\usage{
\method{print}{mcmod.state}(x, ...)
}
\arguments{
\item{x}{the state object}

\item{...}{addition args not used but for method matching}
}
\value{
Prints overview information from state
}
\description{
Interal print method
}
