% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-handler.R
\name{print.mcmodel_result}
\alias{print.mcmodel_result}
\title{Print Method for outputs}
\usage{
\method{print}{mcmodel_result}(x, ...)
}
\arguments{
\item{x}{mcmodel result}

\item{...}{dots to print not used}
}
\value{
Print details of out
}
\description{
Print Method for outputs
}
