% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quantstudio.R
\name{quantstudio-access}
\alias{quantstudio-access}
\alias{get_quantstudio_run_time}
\alias{get_quantstudio_melting_curve}
\alias{get_quantstudio_amplication}
\alias{get_quantstudio_raw}
\alias{get_quantstudio_result}
\title{Access to QuantStudioRaw class object}
\usage{
get_quantstudio_run_time(x)

get_quantstudio_melting_curve(x)

get_quantstudio_amplication(x)

get_quantstudio_raw(x)

get_quantstudio_result(x)
}
\arguments{
\item{x}{QuantStudioRaw object}
}
\value{
a tibble or a vector
}
\description{
\itemize{
\item \code{get_quantstudio_run_time()} - Get run start time
\item \code{get_quantstudio_melting_curve()} - Get melting curve
\item \code{get_quantstudio_amplication()} - Get amplication data
\item \code{get_quantstudio_raw()} - Get raw data
\item \code{get_quantstudio_result()} - Get analysis results
}
}
