% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_quantstudio.R
\name{quantstudio-plot}
\alias{quantstudio-plot}
\alias{plot_quantstudio_melting_curve}
\alias{plot_quantstudio_amplification_curve}
\title{Plot real-time PCR melting/amplification curve}
\usage{
plot_quantstudio_melting_curve(
  data,
  y = c("derivative", "fluorescence"),
  show_tm = FALSE,
  tm_nums = 1
)

plot_quantstudio_amplification_curve(data, y = c("rn", "delta_rn"))
}
\arguments{
\item{data}{plot data}

\item{y}{mapping of y}

\item{show_tm}{if TRUE will plot the TM on melting curve}

\item{tm_nums}{how many TM, default is 1. But can be 2 if have two PCR products.}
}
\value{
a ggplot object
}
\description{
Plot real-time PCR melting/amplification curve
}
