% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantstudioraw-class.R
\docType{class}
\name{QuantStudioRaw-class}
\alias{QuantStudioRaw-class}
\alias{qs_plot}
\alias{print}
\alias{print.QuantStudioRaw}
\title{Methods for QuantStudioRaw class object}
\usage{
qs_plot(x, well_position = NULL)

\method{print}{QuantStudioRaw}(x, ...)
}
\arguments{
\item{x}{a QuantStudioRaw object}

\item{well_position}{specify the well position. If NULL, all wells are included.}

\item{...}{
  Arguments passed on to \code{\link[=print]{print}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
\strong{Input}

The QuantStudioRaw object is a S3 class object by \code{read_quantstudio()} function.

\strong{Output}
\itemize{
\item \code{qs_plot()} - plot object
\item \code{print()} - Print user-friendly information
}
}
