% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_quantstudio.R
\name{read_quantstudio}
\alias{read_quantstudio}
\title{Read all values in QuantStudio export}
\usage{
read_quantstudio(file)
}
\arguments{
\item{file}{all in one text file (*.txt.gz)}
}
\value{
a list of table, QuantStudioRaw class object
}
\description{
Read all values in QuantStudio export
}
