% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rhat}
\alias{rhat}
\alias{gelman_rubin_psrf}
\alias{vehtari_split_psrf}
\title{Potential Scale Reduction Factor (Rhat)}
\usage{
gelman_rubin_psrf(chain_list)

vehtari_split_psrf(chains, rank_norm = TRUE)
}
\arguments{
\item{chain_list}{a list of MCMC Chains}

\item{chains}{either a list of chains or vector of one chain (which will be split)}

\item{rank_norm}{boolean for rank normalization}
}
\value{
numeric score of rhat

vector of rhat score(s)
}
\description{
Functions to calculate rhat. The default used in \link{mcmc_run} is the
rank normalized, split chain R-hat described by Vehtari et al (2021). However,
there is also the ability to calculate the classic Gelman-Rubin Diagnostic.

Classic approach which REQUIRES a chain list with equal length chains.
This is not as built out as the veharti approach and thus cannot support multidimensional output

Based on Vehtari et al 2021; 10.1214/20-BA1221. This version supports both vector chains or multi-dimension arrays (assuming iterations along first axis)
This script is a wrapper which calls \link{vsp_internal}
}
\examples{
n = 1000
gelman_rubin_psrf(list(rnorm(n), rnorm(n))) # should be close to 1
gelman_rubin_psrf(list(rnorm(n), rnorm(n, 100))) #should be large


n = 1000
vehtari_split_psrf(rnorm(n))
vehtari_split_psrf(list(rnorm(n), rnorm(n)), rank_norm = FALSE)

}
\references{
Vehtari A. Gelman A. Simpson D. Carpenter B. B ̈urkner P-C. 2021.
Rank-Normalization, Folding, and Localization:  An Improved ̂R for Assessing Convergence of  MCMC.
Bayesian Analysis 16(2): 667-718. 10.1214/20-BA1221.
Gelman, A. and Rubin, D. B. (1992).
“Inference from iterative simulation using multiple sequences (with discussion).”
Statistical Science, 7(4): 457–511. 667, 668, 671, 675
}
