% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{run_burn}
\alias{run_burn}
\title{Intenral burn in loop}
\usage{
run_burn(state)
}
\arguments{
\item{state}{mcmodel state object}
}
\value{
state after burnin
}
\description{
Intenral burn in loop
}
\keyword{internal}
