% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{run_samples}
\alias{run_samples}
\title{Intenral burn in loop}
\usage{
run_samples(state, n_steps)
}
\arguments{
\item{state}{mcmodel state object}
}
\value{
state at process
}
\description{
Intenral burn in loop
}
\keyword{internal}
