% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-engines.R
\name{simple_mcmc_run}
\alias{simple_mcmc_run}
\title{MCMC Run Simple}
\usage{
simple_mcmc_run(
  data_list,
  const_list,
  init_list,
  prior_pars,
  update_functions,
  burnin = 5000,
  n_iter = 1e+05,
  thin = 1,
  adapt_intv = 100,
  prop_sd = NULL,
  save_names = NULL,
  derv_quants = NULL,
  derived_functions = NULL,
  adapt_tuner = adapt_tuner_double_exp
)
}
\arguments{
\item{burnin}{length of burnin interval}

\item{n_iter}{total numer of iterations}

\item{thin}{thinning interval}

\item{adapt_intv}{adaptation interval for both proposal tunning and thinnin tuning}
}
\value{
list of samples and info from MCMC
}
\description{
Non-adaptive (except proposal tunning) mcmc run
}
\keyword{internal}
