% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve-interpolation.R
\name{sinc_reconstruction}
\alias{sinc_reconstruction}
\title{Reconstruct signal with sinc method}
\usage{
sinc_reconstruction(xx, yy, xi)
}
\arguments{
\item{xx}{a vector of original independent value}

\item{yy}{a vector of original dependent value at xx}

\item{xi}{a vector of target independent variable values for reconstruction}
}
\value{
a vector
}
\description{
Reconstruct signal with sinc method
}
\examples{
  xx = seq(1, 5, by = 0.5)
  yy = sin(xx)
  plot(xx, yy)
  xi = seq(1, 5, by = 0.1)
  yi = sinc_reconstruction(xx, yy, xi)
  plot(xi, yi)
}
