% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{step_adapt}
\alias{step_adapt}
\title{Tune proposal distribution}
\usage{
step_adapt(state)
}
\arguments{
\item{state}{mcmodel state object}
}
\value{
state object
}
\description{
Tune proposal distribution
}
\keyword{internal}
