% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{step_burn_stream}
\alias{step_burn_stream}
\title{Streaming burnin save}
\usage{
step_burn_stream(state)
}
\arguments{
\item{state}{a mcmodel state list}
}
\value{
a state object
}
\description{
Streaming burnin save
}
\keyword{internal}
