% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{step_check_burn}
\alias{step_check_burn}
\title{Check Burn exit}
\usage{
step_check_burn(state)
}
\arguments{
\item{state}{a mcmodel state}
}
\value{
boolean exit
}
\description{
Check Burn exit
}
\keyword{internal}
