% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{step_update}
\alias{step_update}
\title{Update Step}
\usage{
step_update(state)
}
\arguments{
\item{state}{a mcmodel state list}
}
\value{
a state object
}
\description{
Update Step
}
\keyword{internal}
