% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_viz.R
\name{summarize_chain}
\alias{summarize_chain}
\title{Create Chain Summary}
\usage{
summarize_chain(chain)
}
\arguments{
\item{chain}{a chain of single values}
}
\value{
a vector of summary values, slightly more detailed than summarize.
}
\description{
Create Chain Summary
}
\examples{
summarize_chain(rnorm(10000))

}
