% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-handler.R
\name{summary.mcmodel_result}
\alias{summary.mcmodel_result}
\title{Posterior Summary Print}
\usage{
\method{summary}{mcmodel_result}(object, ...)
}
\arguments{
\item{object}{output from \link{mcmc_run}}

\item{...}{dots not used}
}
\value{
a matrix of summary values
}
\description{
Posterior Summary Print
}
