% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-handler.R
\name{trunacte_samples}
\alias{trunacte_samples}
\title{Truncate to the shortest chain length}
\usage{
trunacte_samples(chain_res, min_length)
}
\description{
for multi-chain runs, chains must be similar length
in order to calculate Rhat. This discards some small number of
points. Future extensions of this package may consider tracking/warning
if chains are substantially different lengths.
}
