% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-handler.R
\name{truncate_samples}
\alias{truncate_samples}
\title{Truncate to the shortest chain length}
\usage{
truncate_samples(chain_res, min_length)
}
\arguments{
\item{chain_res}{result from future loop}

\item{min_length}{the minimum lenght of all chains}
}
\value{
a list of truncated samples
}
\description{
for multi-chain runs, chains must be similar length
in order to calculate Rhat. This discards some small number of
points. Future extensions of this package may consider tracking/warning
if chains are substantially different lengths.
}
\keyword{internal}
