% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vehtari_split_psrf}
\alias{vehtari_split_psrf}
\title{Split Chain, Rank Normalized R-hat}
\usage{
vehtari_split_psrf(chains)
}
\description{
Based on Vehtari et al 2021; 10.1214/20-BA1221
this function only takes either a list or vector for ONE CHAIN
The function can take any size array but will return a flattened value for each
chain along the first axis.
}
