% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vsp_internal}
\alias{vsp_internal}
\title{Interal features for Vehtari function}
\usage{
vsp_internal(chains, rank_norm)
}
\arguments{
\item{chains}{a single chains (e.g., like the gelman rubin)}

\item{rank_norm}{boolean to ranknoramlize for not}
}
\value{
numeric of rhat values
}
\description{
Interal features for Vehtari function
}
\keyword{internal}
