% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{write_block}
\alias{write_block}
\title{Write out block point}
\usage{
write_block(state)
}
\arguments{
\item{state}{a mcmodel state}
}
\value{
a flushed state AND writes a block out
}
\description{
Write out block point
}
\keyword{internal}
