% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-state_fx.R
\name{write_check}
\alias{write_check}
\title{Burn Checkponit}
\usage{
write_check(state)
}
\arguments{
\item{state}{mcmodel state obj}
}
\value{
nothing - just writes current state
}
\description{
Burn Checkponit
}
\keyword{internal}
