% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{geweke_check}
\alias{geweke_check}
\title{Geweke convergence diagnostic}
\usage{
geweke_check(chain)
}
\arguments{
\item{chain}{a single vector of an MCMC chain}
}
\value{
boolean
}
\description{
A comparison of means between the first 10th of a chain and the latter half
The proper Geweke diagnostic uses a z-test of means but I used a t-test
}
\examples{
geweke_check(rnorm(1000))

}
\references{
Geweke, J. 1992. Evaluting the accuracy of sampling-based approaches to the calculations
of posterior moments. Bayesian statistics. 4: 641-649.
}
