% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-handler.R
\name{mcmc_run}
\alias{mcmc_run}
\title{Run MCMC}
\usage{
mcmc_run(
  data_list,
  const_list,
  init_list,
  init_generator = NULL,
  prior_pars,
  update_functions,
  adaptive = FALSE,
  save_names = NULL,
  prop_sd = NULL,
  derv_quants = NULL,
  derived_functions = NULL,
  n_chains = 1,
  n_cores = NULL,
  seeds = NULL,
  adapt_tuner = adapt_tuner_double_exp,
  chain_check = geweke_check,
  ...
)
}
\arguments{
\item{data_list}{list object with non-changing data}

\item{const_list}{list object with relevant indexing used in updates/derived fns}

\item{init_list}{list object with ALL parameters used in update/derived, see \link{mcmc_structure}}

\item{init_generator}{(optional) IF n_chains >1, a list of functions to change init_list points. refer to \link{mcmc_structure}}

\item{prior_pars}{list of prior values used in update functions}

\item{update_functions}{a list of update functions used in blocks for MCMC application}

\item{adaptive}{logical - if true adaptive burnin and thinning will be used. Adaptive tunning occurs in both unless adapt_tuner=no_adapt}

\item{save_names}{(optional) list of parameters in init_list to save out. Defaults to init_list}

\item{prop_sd}{IF any Metropolis-Hastings algos are used with, prop_sd much be provided.}

\item{derv_quants}{Character vector specifying derived quantities to track.}

\item{derived_functions}{List of functions for computing derived quantities.}

\item{n_chains}{number of chains to run (if > 1 will be done in parallel)}

\item{n_cores}{optionally specify number of cores for parallel run. If don't want parallel execution, set n_cores=1}

\item{seeds}{optional vector of seeds (equal to n_chains)}

\item{adapt_tuner}{function for tuning prop_sd in MH applications}

\item{chain_check}{function for assess_burnin if run in adaptive burnin.}

\item{...}{arguments passed to the internal MCMC engines, see}
}
\value{
A list of MCMC output with two items;
\itemize{
\item{samples}{list of posterior samplews}
\item{info}{chain diagnostics}
}
}
\description{
High-level wrapper for running simple or adaptive MCMC chains with
options for single or multichain runs (in parallel). See \code{\link{mcmc_structure}} for full detail
for details and examples on properly constructing runs.
}
\examples{
\dontrun{
mcmc_out <- mcmc_run(
 data_list = my_data,
 const_list = my_consts,
 init_list = my_init,
 prior_pars = my_prior,
 update_functions = update_fns,
 chains = 1
)
}

}
