% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-handler.R
\name{mcmc_run}
\alias{mcmc_run}
\title{Run MCMC}
\usage{
mcmc_run(
  data_list,
  const_list,
  init_list,
  init_generator = NULL,
  prior_pars,
  prop_sd = NULL,
  update_functions,
  save_names = NULL,
  derv_quants = NULL,
  derived_functions = NULL,
  n_iter,
  thin,
  n_burn,
  burn_adapt = FALSE,
  min_burn,
  chkpt_freq = "default",
  work_dir = "./run",
  prop_adapt = 100,
  n_chains = 1,
  n_cores = NULL,
  seeds = NULL,
  log_files = TRUE,
  delete_logs = TRUE,
  ...
)
}
\arguments{
\item{data_list}{list object with non-changing data}

\item{const_list}{list object with relevant indexing used in updates/derived fns}

\item{init_list}{list object with ALL parameters used in update/derived, see \link{mcmc_structures}}

\item{init_generator}{(optional) IF n_chains >1, a list of functions to change init_list points. refer to \link{mcmc_structures}}

\item{prior_pars}{list of prior values used in update functions}

\item{prop_sd}{IF any Metropolis-Hastings algos are used with, prop_sd much be provided.}

\item{update_functions}{a list of update functions used in blocks for MCMC application}

\item{save_names}{(optional) list of parameters in init_list to save out. Defaults to init_list}

\item{derv_quants}{Character vector specifying derived quantities to track.}

\item{derived_functions}{List of functions for computing derived quantities.}

\item{n_iter}{integer for number of interations}

\item{thin}{integer for thinning interval. Default is 1.}

\item{n_burn}{integer for maximum burn-in length}

\item{burn_adapt}{logical to decide to use adaptive burn-in procedure.}

\item{min_burn}{integer used if burn_adapt=TRUE, how long minimum burn in needs to be}

\item{chkpt_freq}{interval of when to write out blocked chunks}

\item{work_dir}{directory to work in for logs and checkpoints}

\item{prop_adapt}{interval for tuning proposal in a MH-example}

\item{n_chains}{number of chains to run (if > 1 will be done in parallel)}

\item{n_cores}{optionally specify number of cores for parallel run. If don't want parallel execution, set n_cores=1}

\item{seeds}{optional vector of seeds (equal to n_chains)}

\item{log_files}{logical, should chain outputs be recorded if run in parallel? Default TRUE}

\item{delete_logs}{logical, delete log files once completed. default is TRUE}

\item{...}{arguments passed to the internal MCMC engines, see \link{mcmc_run_internal}}
}
\value{
A list of MCMC output with two items, samples list and info list
}
\description{
High-level wrapper for running simple or adaptive MCMC chains with
options for single or multichain runs (in parallel). See \code{\link{mcmc_structures}} for full detail
for details and examples on properly constructing runs. Algorithmic development was inspired by Hooten & Hobbs 2019, however the code
allows for substantial flexibility. The structure of data inputs were largely inspired by NIMBLE, but intentially without a bulky object-oriented
structure.
}
\details{
There are several features in this default MCMC algorithm which are worth understanding to correct set parameters. First, users
should be familar with how the burn-in procedure is selected. If adaptive-burn in is used, the burn-in is set to be
AT LEAST min_burn and AT MOST n_burn. Thus, if using burn_adapt, it is recommended to increase n_burn to ensure that sufficient
burn-in is provided as the default settings are likely too low in most real cases. If using adaptive burn-in the default exit
is based on \code{\link{geweke_check}}. If MH-steps are used, early exit in adaptive burn-in also requires all acceptance rates to fall
between 0.17 and 0.50. Finally, if working in parallel, it is possible that some chains will end sooner than others in adaptive burn-in
To allow for convergence checks, I automatically trucate longer chains to the shortest one - effectivelly imposing a similar burn-in across samples.

In all cases, there is an adaptive proposal adjustment for MH-updates. The function used to update the prop_sd list is provided with adapt_tuner.
By default this is \code{\link{adapt_tuner_double_exp}}, although user definied functions are possible. To avoid proposal adaptation, set adapt_tuner=no_adapt
While acceptance rate per parameter is not recorded, the overall mean acceptance rate is reported if MH steps are present.

A full working example is available with package source code OR can be produced throgugh following the set-up with \link{mcmc_structures}
}
\examples{
\dontrun{
mcmc_out <- mcmc_run(
 data_list = data_list,
 const_list = const_list,
 init_list = init_list,
 init_generator = init_generator,
 prior_pars = my_prior,
 prop_sd = prop_sd,
 update_functions = update_fns,
 save_names = c('beta', 'sigma'),
 derv_quants = c('p_mse'),
 derived_functions = list('val' = calc_pmse),
 n_iter = 1e4,
 thin = 1,
 min_burn = 5000,
 burn_adapt = TRUE,
 min_burn = 1000,
 n_chains = 10
)
}

}
\references{
Hooten, M.B. and T.J. Hefley. (2019). Bringing Bayesian Models to Life. Chapman and Hall/CRC.
NIMBLE Development Team. 2026. NIMBLE: MCMC, Particle Filtering, and Programmable Hierarchical Modeling.
doi: 10.5281/zenodo.1211190. R package version 1.4.1, https://cran.r-project.org/package=nimble.
}
