% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-handler.R
\name{mcmc_run_internal}
\alias{mcmc_run_internal}
\title{Internal dispatch for a single chain run}
\usage{
mcmc_run_internal(
  data_list,
  const_list,
  init_list,
  prior_pars,
  update_functions,
  adaptive = FALSE,
  save_names = NULL,
  prop_sd = NULL,
  derv_quants = NULL,
  derived_functions = NULL,
  adapt_tuner = adapt_tuner_double_exp,
  chain_check = geweke_check,
  ...
)
}
\arguments{
\item{data_list}{list object with non-changing data}

\item{const_list}{list object with relevant indexing used in updates/derived fns}

\item{init_list}{list object with ALL parameters used in update/derived, see \link{mcmc_structure}}

\item{prior_pars}{list of prior values used in update functions}

\item{update_functions}{a list of update functions used in blocks for MCMC application}

\item{adaptive}{logical - if true adaptive burnin and thinning will be used. Adaptive tunning occurs in both unless adapt_tuner=no_adapt}

\item{save_names}{(optional) list of parameters in init_list to save out. Defaults to init_list}

\item{prop_sd}{IF any Metropolis-Hastings algos are used with, prop_sd much be provided.}

\item{derv_quants}{Character vector specifying derived quantities to track.}

\item{derived_functions}{List of functions for computing derived quantities.}

\item{adapt_tuner}{function for tuning prop_sd in MH applications}

\item{chain_check}{function for assess_burnin if run in adaptive burnin.}

\item{...}{additional args passed to SEE LINK}
}
\description{
Internal dispatch for a single chain run
}
