% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmodel-package.R
\docType{package}
\name{mcmodel-package}
\alias{mcmodel}
\alias{mcmodel-package}
\title{mcmodel: Preprocess, modelling, and visualization of melting curve data}
\description{
This package aims to standardize the preprocessing and modeling prediction process for melting curve data. It currently has two kinds of functions. The first is to implement the reading and storage of melting curves data. It defined a QuantStudioRaw class object to store complete RT-PCR experimental data and a MeltingCurve class object to store complete melting curve data. Based on the MeltingCurve object, we implemented methods for filtering, interpolation, signal processing, and the preparation of training dataset that can be piped to modelling process. Filtering can filter raw data based on sample position, temperature, etc. Interpolation can convert high-resolution melting curves into a form with reduced dimensions and remove temperature biases in fluorescence signal reading during the experiment. Signal processing can detect the features of the melting curve peaks, which has been used in single-species modeling. Data preparation transformed a MeltingCurve object into a data frame. With this data frame, melting curve data can be pipe into R (or Python) machine learning frameworks for modeling and prediction. Correspondingly, methods for training, modeling, and prediction are also integrated into such a package.
}
\author{
\strong{Maintainer}: Chun-Hui Gao \email{gaospecial@gmail.com} (\href{https://orcid.org/0000-0002-1445-7939}{ORCID})

}
\keyword{internal}
